+++ main.py
@@ -1,0 +1,4 @@
+from plugin.database import android_gen
+
+if __name__ == "__main__":
+    android_gen()
